refactored graph package into more files
