--- src/graph/package.go
@@ -1,2 +1,0 @@
-// Package graph includes a graph data structure and some graph utilities
-package graph
