+++ src/graph/partition.go
@@ -1,0 +1,5 @@
+package graph
+
+func (g *Graph) Partition() {
+
+}
